refactor: component organization
