+++ app/[locale]/collectibles/_components/Collectibles/loading.tsx
@@ -1,0 +1,10 @@
+import { Skeleton } from "@/components/ui/skeleton"
+
+const Loading = () => (
+  <div className="flex w-full flex-col gap-x-8 gap-y-6 xl:flex-row">
+    <Skeleton className="h-64 flex-1" />
+    <Skeleton className="h-64 flex-[3]" />
+  </div>
+)
+
+export default Loading
