+++ app/[locale]/collectibles/_components/CollectiblesContent/loading.tsx
@@ -1,5 +1,10 @@
-// TODO: Loading placeholder
+// import { Skeleton } from "@/components/ui/skeleton"
 
-const Loading = () => <div />
+const Loading = () => (
+  <div className="flex w-full flex-col gap-x-8 gap-y-6 bg-black xl:flex-row" />
+  //   <Skeleton className="h-64 flex-1" />
+  //   <Skeleton className="h-64 flex-[3]" />
+  // </div>
+)
 
 export default Loading
