+++ src/intl/en/page-collectibles.json
@@ -1,6 +1,7 @@
 {
   "page-collectibles-hero-title": "Badges",
   "page-collectibles-hero-header": "Ethereum.org Collectibles",
+  "page-collectibles-contributor-img-alt": "Two contributors chatting",
   "page-collectibles-hero-description": "Prove that you worked on ethereum.org, onchain.",
   "page-collectibles-improve-title": "Improve ethereum.org",
   "page-collectibles-improve-desc-1": "Earn unique NFTs by helping maintain and expand ethereum.org website. These badges <strong>recognize your participation onchain</strong>.",
